@@ -1,7 +1,7 @@
-FactoryGirl.define do
+FactoryBot.define do
   factory :schedule do
-    day 'Sunday'
-    location 'Dillard Drive Elementary'
-    time '12:00:00'
+    day { 'Sunday' }
+    location { 'Dillard Drive Elementary' }
+    time { '12:00:00' }
   end
 end
