@@ -1,10 +1,10 @@
-FactoryGirl.define do
+FactoryBot.define do
   factory :user do
-    username 'brandan'
-    email 'brandan@example.com'
-    password 'secret'
+    username { 'brandan' }
+    email { 'brandan@example.com' }
+    password { 'secret' }
     password_confirmation { password }
-    first_name 'Brandan'
-    last_name 'Lennox'
+    first_name { 'Brandan' }
+    last_name { 'Lennox' }
   end
 end
