@@ -9,12 +9,12 @@ Bundler.require(*Rails.groups)
 module FoxyLadyUltimate
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
-    config.load_defaults 5.1
+    config.load_defaults 5.2
 
     # Settings in config/environments/* take precedence over those specified here.
-    # Application configuration should go into files in config/initializers
-    # -- all .rb files in that directory are automatically loaded.
-
+    # Application configuration can go into files in config/initializers
+    # -- all .rb files in that directory are automatically loaded after loading
+    # the framework and any gems in your application.
     config.time_zone = 'Eastern Time (US & Canada)'
     config.active_record.schema_format = :sql
     config.active_job.queue_adapter = :que
