@@ -1,6 +1,6 @@
 RSpec.describe PlayerDecorator do
   describe '#attendance' do
-    let(:player)    { FactoryGirl.create(:player) }
+    let(:player)    { create(:player) }
     let(:decorator) { player.decorate }
 
     context 'when the player has never played' do
@@ -11,7 +11,7 @@ RSpec.describe PlayerDecorator do
     end
 
     context 'when the player has played a few games' do
-      let(:games) { FactoryGirl.create_list(:game, 3, starts_at: 1.week.ago) }
+      let(:games) { create_list(:game, 3, starts_at: 1.week.ago) }
 
       before do
         games.each { |game| game.respond(player, true) }
