@@ -1,6 +1,6 @@
 RSpec.describe Game do
   describe '::seed' do
-    let(:schedule) { FactoryGirl.build(:schedule) }
+    let(:schedule) { build(:schedule) }
     before { expect(Game).to receive(:schedule).at_least(:once).and_return(schedule) }
     before { Game.delete_all }
 
@@ -14,8 +14,8 @@ RSpec.describe Game do
   end
 
   describe '#respond' do
-    let(:player) { FactoryGirl.create(:player) }
-    let(:game)   { FactoryGirl.create(:game) }
+    let(:player) { create(:player) }
+    let(:game)   { create(:game) }
 
     context 'for a positive response' do
       before { game.respond(player, true) }
@@ -48,34 +48,34 @@ RSpec.describe Game do
 
   describe '#upcoming?' do
     context 'when the game is in the past' do
-      subject { FactoryGirl.build(:game, starts_at: 1.day.ago) }
+      subject { build(:game, starts_at: 1.day.ago) }
       it { is_expected.not_to be_upcoming }
     end
 
     context 'when the game is in the future' do
-      subject { FactoryGirl.build(:game, starts_at: 1.day.from_now) }
+      subject { build(:game, starts_at: 1.day.from_now) }
       it { is_expected.to be_upcoming }
     end
   end
 
   describe '#on?' do
     context 'when the game is canceled' do
-      subject { FactoryGirl.build(:game, canceled: true) }
+      subject { build(:game, canceled: true) }
       it { is_expected.not_to be_on }
     end
 
     context 'when the game is not canceled' do
-      subject { FactoryGirl.build(:game, canceled: false) }
+      subject { build(:game, canceled: false) }
       it { is_expected.to be_on }
     end
   end
 
   describe '#default_location?' do
-    let(:schedule) { FactoryGirl.build(:schedule) }
+    let(:schedule) { build(:schedule) }
     before { expect(Schedule).to receive(:instance).and_return(schedule) }
 
     context 'when the game is at the normal location' do
-      subject { FactoryGirl.build(:game, location: schedule.location) }
+      subject { build(:game, location: schedule.location) }
       it { is_expected.to be_default_location }
     end
   end
