@@ -1,5 +1,5 @@
 module SessionHelper
   def session_user
-    @session_user ||= FactoryGirl.create(:user)
+    @session_user ||= create(:user)
   end
 end
