@@ -1,4 +1,6 @@
-RSpec.feature 'Home page' do
+require 'rails_helper'
+
+RSpec.describe 'Home page' do
   it 'shows a heading' do
     visit root_url
     expect(page).to have_content('Next Game')
@@ -12,11 +14,11 @@ RSpec.feature 'Home page' do
   end
 
   context 'when a game is scheduled' do
-    given!(:game) { FactoryGirl.create(:game, starts_at: Time.now + 1.day) }
-    given!(:brandan) { FactoryGirl.create(:player, first_name: 'Brandan', last_name: 'Lennox') }
-    given!(:lindsay) { FactoryGirl.create(:player, first_name: 'Lindsay', last_name: 'Mooring') }
+    let!(:game) { create(:game, starts_at: Time.now + 1.day) }
+    let!(:brandan) { create(:player, first_name: 'Brandan', last_name: 'Lennox') }
+    let!(:lindsay) { create(:player, first_name: 'Lindsay', last_name: 'Mooring') }
 
-    background do
+    before do
       game.respond(brandan, true)
       game.respond(lindsay, false)
     end
@@ -33,7 +35,7 @@ RSpec.feature 'Home page' do
   end
 
   context 'when a game has been canceled' do
-    given!(:game) { FactoryGirl.create(:game, starts_at: Time.now + 1.day, canceled: true) }
+    let!(:game) { create(:game, starts_at: Time.now + 1.day, canceled: true) }
 
     it 'says as much' do
       visit root_url
